# fasofree-backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/orders": {
      "get": {
        "operationId": "OrdersController_findAllForAdmin",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste globale des commandes"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Tour de contrôle : toutes les commandes avec position live des livreurs (SUPER_ADMIN, ADMIN, SUPPORT)",
        "tags": [
          "Orders"
        ]
      },
      "post": {
        "operationId": "OrdersController_createOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Commande créée avec succès"
          },
          "400": {
            "description": "Données de la commande invalides"
          },
          "401": {
            "description": "Utilisateur non authentifié"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Créer une nouvelle commande",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/quote": {
      "post": {
        "operationId": "OrdersController_quoteOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Devis calculé"
          },
          "400": {
            "description": "Données de devis invalides"
          },
          "401": {
            "description": "Utilisateur non authentifié"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtenir un devis : sous-total, frais de livraison (min 800 FCFA), frais plateforme (100 FCFA) et total",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/my-orders": {
      "get": {
        "operationId": "OrdersController_getMyOrders",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Liste des commandes récupérée"
          },
          "401": {
            "description": "Utilisateur non authentifié"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Lister les commandes du client connecté",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{id}": {
      "get": {
        "operationId": "OrdersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Détail de la commande"
          },
          "401": {
            "description": "Utilisateur non authentifié ou rôle manquant"
          },
          "404": {
            "description": "Commande introuvable"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtenir le détail d'une commande par son ID",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{id}/tracking": {
      "get": {
        "operationId": "OrdersController_getOrderTracking",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Données de suivi récupérées"
          },
          "401": {
            "description": "Utilisateur non authentifié ou rôle manquant"
          },
          "404": {
            "description": "Commande introuvable"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Suivi live : statut, position GPS du livreur, tracé du parcours et ETA",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{id}/status": {
      "patch": {
        "operationId": "OrdersController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Statut mis à jour avec succès"
          },
          "400": {
            "description": "Statut invalide"
          },
          "401": {
            "description": "Utilisateur non authentifié"
          },
          "404": {
            "description": "Commande introuvable"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Mettre à jour le statut d'une commande",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{id}/driver-validate": {
      "post": {
        "operationId": "OrdersController_driverValidate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverValidateDeliveryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Livraison marquée - en attente du Code PIN du client"
          },
          "403": {
            "description": "Non autorisé (pas le bon livreur)"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Le livreur/coursier confirme avoir effectué la livraison",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{id}/client-validate": {
      "post": {
        "operationId": "OrdersController_clientValidate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientValidateDeliveryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Commande complétée avec succès !"
          },
          "400": {
            "description": "Code PIN invalide"
          },
          "403": {
            "description": "Cette commande ne vous appartient pas"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Le client confirme la réception avec son Code PIN à 4 chiffres",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{id}/dispute": {
      "post": {
        "operationId": "OrdersController_disputeOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisputeOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Litige ouvert - commande gelée"
          },
          "400": {
            "description": "Statut incompatible pour un litige"
          },
          "403": {
            "description": "Cette commande ne vous appartient pas"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Le client ouvre un litige sur une commande livrée",
        "tags": [
          "Orders"
        ]
      }
    },
...
    "/api/v1/disputes/orders/{orderId}": {
      "post": {
        "operationId": "DisputesController_open",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDisputeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Ouvrir un litige sur une commande livrée",
        "tags": [
          "Disputes"
        ]
      }
    },
...
    "/api/v1/reviews/order/{orderId}": {
      "get": {
        "operationId": "ReviewsController_getReviewsForOrder",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des avis pour cette commande"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtenir les avis d'une commande spécifique",
        "tags": [
          "Reviews"
        ]
      }
    },
...
  },
  "info": {
    "title": "FasoFree API",
    "description": "Spécification OpenAPI officielle du backend FasoFree - Marketplace & Livraison à Ouagadougou",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1",
      "description": "Serveur courant"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Saisissez votre token JWT",
        "in": "header"
      }
    },
    "schemas": {
...
      "P2PLocationDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "description": "Adresse complète",
            "example": "123 Rue Principale, Ouagadougou"
          },
          "latitude": {
            "type": "number",
            "description": "Coordonnée GPS Latitude",
            "example": 12.3714
          },
          "longitude": {
            "type": "number",
            "description": "Coordonnée GPS Longitude",
            "example": -1.5197
          },
          "contactName": {
            "type": "string",
            "description": "Nom du contact",
            "example": "Jean Dupont"
          },
          "contactPhone": {
            "type": "string",
            "description": "Téléphone du contact",
            "example": "+22670123456"
          },
          "instructions": {
            "type": "string",
            "description": "Instructions spécifiques",
            "example": "Sonnette à gauche"
          }
        },
        "required": [
          "address",
          "latitude",
          "longitude",
          "contactName",
          "contactPhone"
        ]
      },
      "PackageDimensionsDto": {
        "type": "object",
        "properties": {
          "length": {
            "type": "number",
            "description": "Longueur en cm",
            "example": 30
          },
          "width": {
            "type": "number",
            "description": "Largeur en cm",
            "example": 20
          },
          "height": {
            "type": "number",
            "description": "Hauteur en cm",
            "example": 15
          }
        }
      },
      "PackageDetailsDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "description": "Description du colis",
            "example": "Clés et documents importants"
          },
          "isFragile": {
            "type": "boolean",
            "description": "Le colis est fragile",
            "example": false
          },
          "weight": {
            "type": "number",
            "description": "Poids du colis en kg",
            "example": 0.5
          },
          "dimensions": {
            "description": "Dimensions du colis",
            "allOf": [
              {
                "$ref": "#/components/schemas/PackageDimensionsDto"
              }
            ]
          }
        }
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "businessId": {
            "type": "string",
            "description": "Identifiant UUID unique de la boutique / du commerce (optionnel pour P2P)",
            "format": "uuid",
            "example": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11"
          },
          "totalAmount": {
            "type": "number",
            "description": "Montant total de la commande en FCFA (XOF)",
            "example": 5000,
            "minimum": 0
          },
          "deliveryLatitude": {
            "type": "number",
            "description": "Coordonnée GPS Latitude de livraison à Ouagadougou",
            "example": 12.3714
          },
          "deliveryLongitude": {
            "type": "number",
            "description": "Coordonnée GPS Longitude de livraison à Ouagadougou",
            "example": -1.5197
          },
          "orderType": {
            "type": "string",
            "description": "Type de commande (ex: LIVRAISON, EMPORTER)",
            "enum": [
              "MERCHANT",
              "P2P_DELIVERY",
              "DELIVERY",
              "PICKUP",
              "RIDE",
              "EXPRESS"
            ]
          },
          "fulfillmentType": {
            "type": "string",
            "description": "Type de fulfillment (DELIVERY, PICKUP, DINE_IN)",
            "enum": [
              "DELIVERY",
              "PICKUP",
              "DINE_IN"
            ]
          },
          "fulfillmentDetails": {
            "type": "object",
            "description": "Détails de fulfillment (tableNumber, reservationTime, numberOfGuests, notes)",
            "example": {
              "tableNumber": "5",
              "numberOfGuests": 2
            }
          },
          "deliveryFee": {
            "type": "number",
            "description": "Frais de livraison appliqués en FCFA (XOF)",
            "example": 1000,
            "minimum": 0
          },
          "promoCode": {
            "type": "string",
            "description": "Code promotionnel contextuel appliqué à la commande",
            "example": "BIENVENUE10"
          },
          "pickupLocation": {
            "description": "Lieu de ramassage pour livraison P2P",
            "allOf": [
              {
                "$ref": "#/components/schemas/P2PLocationDto"
              }
            ]
          },
          "dropoffLocation": {
            "description": "Lieu de livraison pour livraison P2P",
            "allOf": [
              {
                "$ref": "#/components/schemas/P2PLocationDto"
              }
            ]
          },
          "packageDetails": {
            "description": "Détails du colis pour livraison P2P",
            "allOf": [
              {
                "$ref": "#/components/schemas/PackageDetailsDto"
              }
            ]
          }
        },
        "required": [
          "totalAmount",
          "orderType"
        ]
      },
      "QuoteOrderDto": {
        "type": "object",
        "properties": {
          "orderType": {
            "type": "string",
            "description": "Type de commande à deviser : DELIVERY / MERCHANT (commerce) ou P2P_DELIVERY (course)",
            "enum": [
              "MERCHANT",
              "P2P_DELIVERY",
              "DELIVERY",
              "PICKUP",
              "RIDE",
              "EXPRESS"
            ],
            "example": "DELIVERY"
          },
          "businessId": {
            "type": "string",
            "description": "Identifiant UUID de la boutique (optionnel si businessLatitude/businessLongitude fournis, inutile pour P2P)",
            "format": "uuid"
          },
          "deliveryLatitude": {
            "type": "number",
            "description": "Coordonnée GPS Latitude de livraison",
            "example": 12.3714
          },
          "deliveryLongitude": {
            "type": "number",
            "description": "Coordonnée GPS Longitude de livraison",
            "example": -1.5197
          },
          "businessLatitude": {
            "type": "number",
            "description": "Coordonnée GPS Latitude de la boutique (fallback si pas de businessId en base)",
            "example": 12.3665
          },
          "businessLongitude": {
            "type": "number",
            "description": "Coordonnée GPS Longitude de la boutique (fallback si pas de businessId en base)",
            "example": -1.4807
          },
          "subtotal": {
            "type": "number",
            "description": "Sous-total des articles en FCFA (0 pour une course P2P). Le backend utilise ce montant dans le devis mais recompte TOUJOURS au POST /orders.",
            "example": 7500,
            "minimum": 0
          },
          "pickupLocation": {
            "description": "Lieu de ramassage (obligatoire pour P2P_DELIVERY)",
            "allOf": [
              {
                "$ref": "#/components/schemas/P2PLocationDto"
              }
            ]
          },
          "dropoffLocation": {
            "description": "Lieu de livraison (obligatoire pour P2P_DELIVERY)",
            "allOf": [
              {
                "$ref": "#/components/schemas/P2PLocationDto"
              }
            ]
          }
        },
        "required": [
          "orderType"
        ]
      },
      "UpdateOrderStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Nouveau statut à appliquer à la commande",
            "enum": [
              "PENDING",
              "PAID",
              "IN_PREPARATION",
              "PROCESSING",
              "DELIVERED_PENDING_CONFIRMATION",
              "DELIVERED",
              "COMPLETED",
              "CANCELLED",
              "FAILED",
              "DISPUTED",
              "REFUNDED"
            ],
            "example": "PAID"
          }
        },
        "required": [
          "status"
        ]
      },
      "DriverValidateDeliveryDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "description": "Note optionnelle du livreur (ex: \"Remis au gardien\")",
            "example": "Remis au gardien de l'immeuble"
          }
        }
      },
      "ClientValidateDeliveryDto": {
        "type": "object",
        "properties": {
          "pinCode": {
            "type": "string",
            "description": "Code PIN à 4 chiffres fourni au client lors de la création de la commande",
            "example": "4829",
            "minLength": 4,
            "maxLength": 4
          }
        },
        "required": [
          "pinCode"
        ]
      },
      "DisputeOrderDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "description": "Raison du litige",
            "example": "Commande non reçue malgré la confirmation du livreur"
          }
        },
        "required": [
          "reason"
        ]
      },
...
      "CreateDisputeDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Commande non reçue malgré la confirmation du livreur."
          },
          "attachments": {
            "description": "Clés S3 déjà téléversées via le module Upload.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "reason"
        ]
      },
...
    }
  }
}